# sstable: perf improvement for ResetForReuse